requirements.txt
flask
oracledb
zeep
python-dotenv
Pillow
requests
Flask-Cors